Implement ConnectionInstance in Typescript
